backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
celery==5.4.0
redis==5.1.1
rembg==2.0.57
Pillow==10.4.0
aiofiles==24.1.0
python-multipart==0.0.12
